requirements.txt
argparse
dkimpy
email-to
pydig
python-whois
secure-smtplib